Followed the linting instructions as much as possible
