--- SphereSurfaceChargeDensity/UnitTestingModule.py
@@ -1,27 +1,0 @@
-"""
-This module is to test for the surface charge density of a conducting spherical shell of radius (R) held at constant potential (V).
-"""
-
-# Import the function from the module where it's defined
-from PhysicsModule import *
-
-# Permittivity of free space
-Epsilon_0 = 8.854*(10**(-12))
-
-# Define some test cases
-def test_sigma():
-    """
-    A function that tests for different cases two of which are limiting scenarios (namely, the last two cases [for V = 0, and for large R])
-    Parameters: None
-    Returns: either returns messages for the first encountered failed test or prints "All tests passed"
-    """
-    radii = [1.0, 2.0, 1.0, 1*(10**6)]
-    potentials = [100.0, 200.0, 0.0, 100.0]
-    calculated_results = map(sigma, radii, potentials) # the map function
-    expected_results = [Epsilon_0 * (potentials[i] / radii[i]) for i in range(len(potentials))] # List-comprehensions
-    for i in range(len(potentials)):
-        assert next(calculated_results) == expected_results[i], f"Test case {i} failed"
-    print("All tests passed")
-    
-# Call the test function
-test_sigma()
