+++ SphereSurfaceChargeDensity/math_ops_module.py
@@ -1,21 +1,27 @@
-""" This Module is to define the general mathematical operations to be used by the other modules """
+"""
+This Module is to define the general mathematical operations to be used by the other modules
+"""
 
 def div(numerator, denominator):
     """
     Returns the result of dividing two numbers
+    
     Parameters:
     numerator, denominator: input numbers
+    
     Returns:
-    number: The result of dividing the first number by the second number utilizing the lambda expression
+    number: The result of dividing the first number by the second
     """
     return numerator / denominator
 
 def mult(x, y):
     """
     Returns the result of multiplying two numbers
+    
     Parameters:
     x, y: input numbers
+    
     Returns:
-    number: The result of multiplying the two numbers utilizing the lambda expression
+    number: The result of multiplying the two numbers
     """
     return x * y
